9/8 post session management
게시물 생성시 로그인,글쓴이 확인

+++ src/main/java/com/spring/instagram/create/PostService.java
@@ -1,8 +1,10 @@
 package com.spring.instagram.create;
 
+import com.spring.instagram.login.UserInfo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import javax.annotation.Resource;
 import java.util.List;
 import java.util.Optional;
 
@@ -11,9 +13,13 @@ import java.util.Optional;
 public class PostService {
     private PostRepository postRepository;
 
+    @Resource
+    private UserInfo userInfo;
+
     @Autowired
-    public PostService(PostRepository postRepository){
+    public PostService(PostRepository postRepository,UserInfo userInfo){
         this.postRepository=postRepository;
+        this.userInfo=userInfo;
     }
     @GetMapping
     public List<Post> postList() {
@@ -23,11 +29,16 @@ public class PostService {
     @PostMapping
     public String createPost(@RequestBody Post post){
         try{
-            this.postRepository.save(post);
-            return "create Post";
+            if(userInfo.getUserNm()!=null && userInfo.getUserId()==post.getWrite_by().getUserId()) {
+                this.postRepository.save(post);
+                return "create Post";
+            }
+            else{
+                return "please login";
+            }
         }
         catch (Exception e){
-            return "Fail";
+            return "fail";
         }
 
     }
@@ -35,7 +46,7 @@ public class PostService {
     public String updatePost(@RequestBody Post post) {
         try {
             Optional<Post> postOptional = this.postRepository.findById(post.getId());
-            if (postOptional.isPresent()) {
+            if (postOptional.isPresent() &&userInfo.getUserNm()!=null && userInfo.getUserId()==post.getWrite_by().getUserId()) {
                 Post updatePost = postOptional.get();
                 updatePost.setBody(post.getBody());
                 updatePost.setWrite_by(post.getWrite_by());
@@ -55,7 +66,7 @@ public class PostService {
     public String deletePost(@PathVariable Long id){
         try {
             Optional<Post> postOptional = this.postRepository.findById(id);
-            if (postOptional.isPresent()) {
+            if (postOptional.isPresent()&& userInfo.getUserNm()!=null && userInfo.getUserId()==postOptional.get().getWrite_by().getUserId()) {
                 this.postRepository.delete(postOptional.get());
             } else {
                 return "delete Fail";
